requirements.txt
flask
torch
diffusers
transformers
accelerate
pillow
